storage: move `DelimitedValueReader` to its own module
Pure code movement and fixing up use statements.
